% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_benchmark.R
\name{ZZ_benchmark}
\alias{ZZ_benchmark}
\title{Create ZZ Benchmark Return Series}
\usage{
ZZ_benchmark(
  start,
  end,
  frequency = c("daily", "weekly", "monthly"),
  type = c("simple", "log"),
  cum = FALSE
)
}
\arguments{
\item{start}{Start date (Date or character in YYYY-MM-DD format).}

\item{end}{End date (Date or character in YYYY-MM-DD format).}

\item{frequency}{Defines the frequency of the return series ("daily", "weekly", "monthly")}

\item{type}{Defines if simple or log returns should be returned}

\item{cum}{Activates cumulative returns}
}
\value{
A data frame with columns: \code{date} and \code{weighted_return}.
}
\description{
Downloads price data from the MSCI World and MSCI Emerging Markets ETF from Yahoo Finance and applies
a predefined 60/40 weighting to calculate a weighted portfolio return.
}
\examples{
\dontrun{
  # Example:
  ZZ_benchmark(
    start = "2020-01-01",
    end = Sys.Date(),
    frequency = "weekly",
    type = "simple",
    cum = TRUE
  )
}
}
