% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_corrplot.R
\name{ZZ_corrplot}
\alias{ZZ_corrplot}
\title{Download Portfolio Data and create Correlation Analysis}
\usage{
ZZ_corrplot(
  df = NULL,
  frequency = c("daily", "weekly", "monthly"),
  type = c("simple", "log"),
  fuzzy = TRUE,
  max_days = 5
)
}
\arguments{
\item{df}{Dataframe with the return series including date column (Optional).}

\item{frequency}{Defines the frequency of the return series ("daily", "weekly", "monthly").}

\item{type}{Defines if simple or log returns should be returned.}

\item{fuzzy}{logical; if TRUE, nearest-date join is used for `df`.}

\item{max_days}{integer; maximum day distance for nearest-date join.}
}
\value{
A corrplot with correlation matrix visualised.
}
\description{
Downloads price data for all portfolio holdings from Yahoo Finance and conducts
a correlation analysis with the help of corrplot.
It plots the correlation plot and saves the correlation matrix.
}
\examples{
\dontrun{
  # Example:
  ZZ_corrplot(
    df = strat_ret,
    frequency = "weekly",
    type = "simple",
  )
}
}
