% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_remove_na.R
\name{ZZ_remove_na}
\alias{ZZ_remove_na}
\title{Data Cleaning of data frames}
\usage{
ZZ_remove_na(
  df,
  row_thresh = 0.5,
  col_thresh = 0.5,
  interpolate = FALSE,
  protect = "date"
)
}
\arguments{
\item{df}{Dataframe with the data.}

\item{row_thresh}{Defines the threshold (in proportion) for deleting a row (base is 0.5).}

\item{col_thresh}{Defines the threshold (in proportion) for deleting a column (base is 0.5).}

\item{interpolate}{Optionally, linearly interpolate missing data points}

\item{protect}{Character vector of column names to keep untouched (preset to "date").}
}
\value{
Returns a cleaned data frame for further analysis
}
\description{
Delete all columns and or rows of a data frame with NAs up to a specified threshold.
Optionally, missing data points can be linearly interpolated.
}
\examples{
\dontrun{
  # Example:
  ZZ_remove_na(
    df = strat_ret,
    row_thresh = 0.5,
    col_thresh = 0.5,
    interpolate = FALSE,
    protect = "date")
}
}
